+++ CHANGELOG.rst
@@ -4,8 +4,9 @@ CHANGELOG
 Next release
 ------------
 
-* Remove unsupported and deprecated Nginx/Pagespeed settings
+* Remove unsupported and deprecated Nginx/Pagespeed settings.
 * Send release and environment tracking info along with Sentry events.
+* Remove unused dependencies.
 
 
 1.11.10.1 (2018-02-02)